the code about recruit question
All the code assignments have been placed in the ForSynX folder (including the basic, backend, and resume sections. All solutions are consolidated into Markdown documents within their respective subfolders, with the code also pasted into the documents. Of course, the original code files are also in the folders). MyTetris is the code for the final game project in the backend track. Unfortunately, only a small portion has been completed.